requirements.txt
Django==3.0.4
gunicorn==20.0.4
django-heroku==0.3.1
requests==2.23.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
django-filter==2.3.0
psycopg2==2.8.4
drf-yasg==1.17.1